Add credential checking for sensitive actions using accessToken

+++ SkyForceWebService/src/main/java/com/skyforce/SkyForceWebService/controller/CustomerController.java
@@ -5,13 +5,11 @@ import com.skyforce.SkyForceWebService.model.Customer;
 import com.skyforce.SkyForceWebService.service.CustomerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.security.NoSuchAlgorithmException;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicLong;
 
 @RestController
 public class CustomerController {
@@ -20,37 +18,25 @@ public class CustomerController {
 
     @Autowired
     CustomerService customerService;
-    private AtomicLong nextId = new AtomicLong();
 
     @GetMapping("/customers")
     public String getAllCustomers() {
         List<Customer> customers = customerService.findAll();
         return JSONConvert.JSONConverter(customers);
     }
-
-    @GetMapping("/customertoken")
-    public String fooo(
-            @RequestHeader(value = "Accept") String accept,
-            @RequestHeader(value = "Accept-Language") String acceptLanguage,
-            @RequestHeader(value = "User-Agent", defaultValue = "foo") String userAgent,
-            @RequestHeader(value = "Authorization") String accessToken
-    ) {
-        System.out.println("accept: " + accept);
-        System.out.println("acceptLanguage: " + acceptLanguage);
-        System.out.println("userAgent: " + userAgent);
-        System.out.println("authentication " + accessToken);
-        return accept + " " + acceptLanguage + " " + userAgent + " " + ValidationController.decryptAccessToken(accessToken);
-    }
+    // pass
 
     @GetMapping("/customer")
     public String getCustomerById(
-            @RequestParam Long customerId) {
+            @RequestParam Long customerId
+    ) {
         Customer customer = customerService.findCustomerById(customerId);
         return JSONConvert.JSONConverter(customer);
     }
+    // pass
 
     @GetMapping("/customer/signin")
-    public String signInCustomer (
+    public String signInCustomer(
             @RequestParam String email,
             @RequestParam String password
     ) throws NoSuchAlgorithmException {
@@ -60,12 +46,8 @@ public class CustomerController {
         }
         return ValidationController.UserSignIn(customer, password);
     }
-    @ResponseStatus(value = HttpStatus.UNAUTHORIZED,
-            reason = "Email or password incorrect")
-    @ExceptionHandler(IllegalArgumentException.class)
-    public void badAuthenticationException() {
+    // pass
 
-    }
 
     @PostMapping("/customer")
     @ResponseStatus(HttpStatus.CREATED)
@@ -77,30 +59,58 @@ public class CustomerController {
         JSONConvert.JSONConverter(customerService.save(customer));
         return ValidationController.getAccessToken(customer.getId(), "CUSTOMER");
     }
+    // pass
 
     @PutMapping("/customer")
     public String updateCustomerById(
             @RequestParam Long customerId,
+            @RequestHeader(value = "Authorization") String accessToken,
             @Valid @RequestBody Customer customer
     ) {
-        Customer oldCustomer = customerService.findCustomerById(customerId);
-        oldCustomer.setFirstName(customer.getFirstName());
-        oldCustomer.setLastName(customer.getLastName());
-        oldCustomer.setGender(customer.getGender());
-        oldCustomer.setDob(customer.getDob());
-        oldCustomer.setAvatar(customer.getAvatar());
-        oldCustomer.setHalaPreference(customer.isHalaPreference());
-        oldCustomer.setVegPreference(customer.isVegPreference());
-        oldCustomer.setRegionalPreference(customer.getRegionalPreference());
-        Customer updatedCustomer = customerService.save(oldCustomer);
-        return JSONConvert.JSONConverter(updatedCustomer);
+        String[] info = ValidationController.decryptAccessToken(accessToken);
+        if (info.length != 2)
+            throw new IllegalArgumentException();
+        if (Long.parseLong(info[0]) == customerId && info[1].equals("CUSTOMER")) {
+            Customer oldCustomer = customerService.findCustomerById(customerId);
+            oldCustomer.setFirstName(customer.getFirstName());
+            oldCustomer.setLastName(customer.getLastName());
+            oldCustomer.setGender(customer.getGender());
+            oldCustomer.setDob(customer.getDob());
+            oldCustomer.setAvatar(customer.getAvatar());
+            oldCustomer.setHalaPreference(customer.isHalaPreference());
+            oldCustomer.setVegPreference(customer.isVegPreference());
+            oldCustomer.setRegionalPreference(customer.getRegionalPreference());
+            customerService.save(oldCustomer);
+            return "OK";
+        }
+        else {
+            throw new IllegalArgumentException();
+        }
     }
+    //pass
 
     @DeleteMapping("/customer")
-    public ResponseEntity<?> deleteCustomer(
-            @RequestParam Long customerId) {
-        Customer customer = customerService.findCustomerById(customerId);
-        customerService.delete(customer);
-        return ResponseEntity.ok().build();
+    public String deleteCustomer(
+            @RequestParam Long customerId,
+            @RequestHeader(value = "Authorization") String accessToken
+    ) {
+        String[] info = ValidationController.decryptAccessToken(accessToken);
+        if (info.length != 2)
+            throw new IllegalArgumentException();
+        if (Long.parseLong(info[0]) == customerId && info[1].equals("CUSTOMER")) {
+            Customer customer = customerService.findCustomerById(customerId);
+            customerService.delete(customer);
+            return "OK";
+        } else {
+            throw new IllegalArgumentException();
+        }
+    }
+    //pass
+
+    @ResponseStatus(value = HttpStatus.UNAUTHORIZED,
+            reason = "Email or password incorrect")
+    @ExceptionHandler(IllegalAccessException.class)
+    public void badAuthenticationException() {
+
     }
 }
